Many, many bug fixes, connection tags, tags-from-tags
New Features:

- Create tags from existing tags.  plc_tag_create_from_tag().
- Connection tags: A new tag type that reports connection state
transitions (up/down/reconnecting) as callback events. Supports
tag-from-tag creation (cloning a connection tag from an existing data
tag) and optional IO events (read/write started/completed).
- Connection state events: Connection status changes are now delivered
as new callback events. Duplicate events are folded.
- Slot 16 support. Some ControlLogix chassis have more slots with slot
indexes up to 16.
- Experimental fiber-based AB and Modbus servers.
- Minor version bump: Reflects the new public API surface.

Bug Fixes (mostly Omron):

- Exponential backoff on Omron reconnection (now same as AB).
- Omron teardown ordering. Caused SISEGV and possible hangs. Possible
fix for [625](https://github.com/libplctag/libplctag/issues/625)
- Race condition on tag removal: Removing a tag could set the terminate
flag on its connection while the refcount was still non-zero, causing a
subsequent data tag to attach to an already-terminating connection.
- Request ref-count race: rc_dec(req) was called before abort_request,
creating a window where the request could be freed while still being
used. Fixed ordering. Possible fix for
[625](https://github.com/libplctag/libplctag/issues/625)
- Request leak race: A race in the path that removes a request from a
tag could cause a leak.
- Spin-block fall-through: Missing rc != PLCTAG_STATUS_OK check could
allow execution to continue incorrectly.
- Wrong pthread clock base: CLOCK_MONOTONIC was used where
CLOCK_REALTIME is required by pthread_cond_timedwait. Also fixed
reversed success/failure status checks in the macOS locking wrappers.
- Integer type / format string issues: "%lld" and (long long) casts
caused warnings and incorrect behavior on 32-bit Linux and musl.

Build / CI Fixes:
  
- 32-bit x86 support (Windows MinGW and Linux).  Edge case fixes.
- MSVC compatibility fix (max_align_t in stddef.h).
- GCC 16+ C++ static linking workaround. Fixes build on Fedora 44
[628](https://github.com/libplctag/libplctag/issues/628)
- Various MinGW/MSYS2 path and shell quoting fixes.
- Modbus server startup race (bind() before SO_REUSEADDR) and
kill/restart timing fixed
  in test scripts.

Thanks to all the users who reported the bugs and other issues!

+++ src/poc/modbus_server/modbus_server.c
@@ -223,7 +223,7 @@ static void print_statistics(server_ctx_t *server) {
     fprintf(stderr, "║        MODBUS SERVER (COROUTINE) PERFORMANCE STATISTICS          ║\n");
     fprintf(stderr, "╠══════════════════════════════════════════════════════════════════╣\n");
     fprintf(stderr, "║ Runtime: %.2f seconds                                            \n", runtime_sec);
-    fprintf(stderr, "║ Total requests: %" PRId64 "                                              \n", (long long)total_reqs);
+    fprintf(stderr, "║ Total requests: %" PRId64 "                                              \n", (int64_t)total_reqs);
     if(runtime_sec > 0) {
         fprintf(stderr, "║ Throughput: %.2f requests/sec                                    \n",
                 (double)total_reqs / (double)runtime_sec);
@@ -232,9 +232,9 @@ static void print_statistics(server_ctx_t *server) {
     fprintf(stderr, "║                       CLIENT STATISTICS                          ║\n");
     fprintf(stderr, "╠══════════════════════════════════════════════════════════════════╣\n");
     fprintf(stderr, "║ Clients connected:    %" PRId64 "                                     \n",
-            (long long)stats->clients_connected);
+            (int64_t)stats->clients_connected);
     fprintf(stderr, "║ Clients disconnected: %" PRId64 "                                     \n",
-            (long long)stats->clients_disconnected);
+            (int64_t)stats->clients_disconnected);
     fprintf(stderr, "╠══════════════════════════════════════════════════════════════════╣\n");
     fprintf(stderr, "║                     RESPONSE TIME SUMMARY                        ║\n");
     fprintf(stderr, "╠══════════════════════════════════════════════════════════════════╣\n");
@@ -246,8 +246,8 @@ static void print_statistics(server_ctx_t *server) {
 
         fprintf(stderr, "║ Average:  %8.2f us                                            \n", mean);
         fprintf(stderr, "║ Std Dev:  %8.2f us                                            \n", stddev);
-        fprintf(stderr, "║ Minimum:  %8lld us                                            \n", (long long)min_time);
-        fprintf(stderr, "║ Maximum:  %8lld us                                            \n", (long long)max_time);
+        fprintf(stderr, "║ Minimum:  %8" PRId64 " us                                            \n", (int64_t)min_time);
+        fprintf(stderr, "║ Maximum:  %8" PRId64 " us                                            \n", (int64_t)max_time);
 
         fprintf(stderr, "╠══════════════════════════════════════════════════════════════╣\n");
         fprintf(stderr, "║                    LATENCY BREAKDOWN (avg)                       ║\n");
@@ -294,7 +294,7 @@ static void print_statistics(server_ctx_t *server) {
         fprintf(stderr, "║  %-12s │", hist_bucket_label(i));
         for(int j = 0; j < bar_len; j++) { fprintf(stderr, "█"); }
         for(int j = bar_len; j < 30; j++) { fprintf(stderr, " "); }
-        fprintf(stderr, "│ %6lld (%5.1f%%)\n", (long long)count, pct);
+        fprintf(stderr, "│ %6" PRId64 " (%5.1f%%)\n", (int64_t)count, pct);
     }
 
     fprintf(stderr, "╚══════════════════════════════════════════════════════════════════╝\n");
